Trait winrt::Abi

pub unsafe trait Abi: Sized {
    type Abi;
    fn abi(&self) -> Self::Abi { ... }

    unsafe fn set_abi(&mut self) -> *mut Self::Abi { ... }

    unsafe fn from_abi(abi: Self::Abi) -> Result<Self> { ... }
}

Provides a generic way of referring to and converting between a Rust object and its WinRT ABI equivalent.

This trait is automatically used by the generated bindings and should not be used directly.

Associated Types

type Abi

Loading content...
Provided methods

fn abi(&self) -> Self::Abi

Casts the Rust object to its ABI type without copying the object.

unsafe fn set_abi(&mut self) -> *mut Self::Abi

Returns a pointer for setting the object's value via an ABI call. This default implemnetation is always correct. Only override if you need to assert something for debugging purposes.

unsafe fn from_abi(abi: Self::Abi) -> Result<Self>

Casts the ABI representation to a Rust object by taking ownership of the bits. This default implementation is correct for all but interfaces.

Loading content...